be/pyproject.toml
[tool.poetry]
name = "travelpicker"
version = "0.1.0"
description = ""
authors = ["GwonilJoo <gwonil3256@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
aiohttp = "^3.9.5"
pyjwt = "^2.8.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
travelpicker = "manage:app"